style(templates): rename templates with color-related names
