+++ globals/index.ts
@@ -1,9 +1,9 @@
 import { type Template } from "@/types";
 
 export const templates: Array<Template> = [
-  "Benjent2023",
-  "Benjent2023Light",
-  "Caillou2023Light",
+  "CottonCandy",
   "OpenResume",
-  "Pharmacy2018",
+  "Pharmacy",
+  "Toothpaste",
+  "Stone",
 ];
