+++ templates/Pharmacy.vue
