+++ templates/Stone.vue
