+++ templates/Toothpaste.vue
